requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.6.0
azure-cognitiveservices-speech==1.28.0
blinker==1.6.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.1.3
et-xmlfile==1.1.0
Flask==2.3.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
openpyxl==3.0.9
pandas==2.0.1
pandas-stubs==2.0.1.230501
PyAudio==0.2.13
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SpeechRecognition==3.10.0
toml==0.10.2
tqdm==4.65.0
types-pytz==2023.3.0.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
